Update lando zsh plugin to work with new version
Lando has removed the registry file, so we must now rely on using the
lando list command. This may cause performance issues in the future
since it has to run the full command every time. Ideally this script
gets updated to cache the output for X amount of time (or another more
precise invalidation method).