Fix snippet tab completion jumping cursor around
If you enter a snippet, but didn't tab to each of it's placeholders and
then went and tried to insert a literal tab somewhere else in the file
(after exiting insert mode from the original snippet) it would jump your
cursor back to any leftover placeholders from the snippet. This fixes
that by "unlinking" the current snippet when leaving insert mode.